patch 8.1.0285: compiler warning for conversion
Problem:    Compiler warning for conversion.
Solution:   Add a type cast. (Mike Williams)

+++ src/ex_getln.c
@@ -580,7 +580,7 @@ may_adjust_incsearch_highlighting(
     {
 	pat = last_search_pattern();
 	skiplen = 0;
-	patlen = STRLEN(pat);
+	patlen = (int)STRLEN(pat);
     }
     else
 	pat = ccline.cmdbuff + skiplen;
